Pricing update
Fluffy pricing is now based on colors (with a tiered system), as well as
a bonus from stats and training.
You can now train a fluffy, training boosts the sell price.
Playing with a fluffy increases Trust, whicih has a small boost to sale
price.

Buying, playing with and training fluffys now take energy.
You start the week with 10 actions, and get that back on the start of a
new week.


Signed-off-by: Eidalac <eidalac@gmail.com>