Event prototype
Mating and birth are now run as events (so will display each on there
own page).
There is now a rejection event for brown foals as a prototype.
- text based on mother's personality.
- option to punish the mother, forcing her to raise the foal with a
chance of giving a "good mama" trait that will prevent rejection of
foals.
- option to ignore the foal, which will feed on poop, be stunted and
unhealthy.
- option to hand rear the foal.
all options need refinment but the basic funtionalty is in place.
various spelling corrections.