Bugfixes
Limited customers per week to 10.
Selling nursing foals should correctly update the mother.
Put an ID check when selling (seems some fluffies don't have a correct
ID which causes a random fluffy to be sold), still looking into the root
cause of that ID error, but a temp fix is in place